Move challenge metadata to existing challenge util
