+++ packages/nextjs/app/challenge/[challengeId]/page.tsx
@@ -4,10 +4,10 @@ import { MDXRemote } from "next-mdx-remote/rsc";
 import rehypeRaw from "rehype-raw";
 import remarkGfm from "remark-gfm";
 import { ArrowTopRightOnSquareIcon } from "@heroicons/react/24/outline";
-import { CHALLENGE_METADATA } from "~~/services/challenges-metadata";
 import { ChallengeId } from "~~/services/database/config/types";
 import { getAllChallenges, getChallengeById } from "~~/services/database/repositories/challenges";
 import { fetchGithubReadme, parseGithubUrl } from "~~/services/github";
+import { CHALLENGE_METADATA } from "~~/utils/challenges";
 import { getMetadata } from "~~/utils/scaffold-eth/getMetadata";
 
 // 6 hours
