+++ packages/nextjs/utils/challenges.ts
@@ -1,7 +1,60 @@
-import { ReviewAction } from "~~/services/database/config/types";
+import { ChallengeId, ReviewAction } from "~~/services/database/config/types";
 
 export const REVIEW_ACTION_BADGE_CLASSES: Record<ReviewAction, string> = {
   [ReviewAction.ACCEPTED]: "badge-success",
   [ReviewAction.REJECTED]: "badge-error dark:bg-[#6E2A2A] dark:text-gray-200",
   [ReviewAction.SUBMITTED]: "badge-warning",
 } as const;
+
+export const CHALLENGE_METADATA: Record<string, { title: string; description: string }> = {
+  [ChallengeId.SIMPLE_NFT_EXAMPLE]: {
+    title: "Learn How to Create an NFT in Solidity",
+    description:
+      "Build your first NFT smart contract in Solidity. Step-by-step tutorial to mint and deploy NFTs on Ethereum using Scaffold-ETH.",
+  },
+  [ChallengeId.DECENTRALIZED_STAKING]: {
+    title: "Build a Decentralized Staking Application in Solidity",
+    description:
+      "Learn to build a decentralized staking application with Solidity. Create smart contracts for coordinating group funding on Ethereum.",
+  },
+  [ChallengeId.TOKEN_VENDOR]: {
+    title: "Create an ERC20 Token and Vendor Contract in Solidity",
+    description:
+      "Learn to create your own ERC20 token and build a decentralized vending machine using Solidity smart contracts on Ethereum.",
+  },
+  [ChallengeId.DICE_GAME]: {
+    title: "Solidity Randomness and Security: Build a Dice Game",
+    description:
+      "Learn about randomness challenges in blockchain and exploit vulnerabilities in a Solidity dice game. Essential security concepts for blockchain developers.",
+  },
+  [ChallengeId.MINIMUM_VIABLE_EXCHANGE]: {
+    title: "How to Build a Decentralized Exchange (DEX) in Solidity",
+    description:
+      "Step-by-step tutorial to build your own DEX with liquidity pools, token swapping, and automated market making in Solidity.",
+  },
+  [ChallengeId.MULTISIG]: {
+    title: "Build a Multisignature Wallet in Solidity",
+    description:
+      "Learn to build a multisignature wallet in Solidity that requires multiple approvals for transactions. Essential security pattern for Ethereum developers.",
+  },
+  [ChallengeId.SVG_NFT]: {
+    title: "Create On-Chain SVG NFTs with Solidity",
+    description:
+      "Learn to generate on-chain SVG graphics and create fully on-chain NFTs using Solidity smart contracts. Advanced NFT development tutorial.",
+  },
+  [ChallengeId.STABLECOINS]: {
+    title: "Build a Stablecoin in Solidity",
+    description:
+      "Learn how to create a stablecoin using Solidity smart contracts on Ethereum. Understand stability mechanisms and price oracles.",
+  },
+  [ChallengeId.PREDICTION_MARKETS]: {
+    title: "Create a Prediction Market in Solidity",
+    description:
+      "Learn to build decentralized prediction markets using Solidity. Understand market creation, betting mechanics, and outcome resolution.",
+  },
+  [ChallengeId.DEPLOY_TO_L2]: {
+    title: "Deploy Solidity Smart Contracts to Layer 2",
+    description:
+      "Learn how to deploy and interact with Solidity smart contracts on Ethereum Layer 2 solutions for better scalability and lower gas costs.",
+  },
+};
